requirements.txt
python-telegram-bot==22.6
mcp==1.26.0
ollama==0.6.1
python-dotenv==1.0.1
httpx==0.28.1
